Only sends notifications if there was a change
We inspect the config event and make sure that this is a change,
before notifying. AWS Config re-checks every resource in the account
every 6 hours when the data is written to S3, we don't want to trigger
notifications on these events.

Also updates the Anghammarad client, because this allows us to
separate the logic for creating messages from the code that actually
sends a message, which makes the following easier.

This change is supported by a refactor to separate the message logic
from the Anghammarad invocation. This means we can test the new
behaviour.

We also remove the Main class because it was going to be some work to
fix it and at this stage it has diverged so much that it there's not
much value in it. The increased test coverage helps here, too.

We could write a new main that generates config event JSON and passes
that to the program so we are actually running something that looks
like the Lambda.

NOTE: this currently refers to an Anghammarad snapshot for the
notification updates. Once the following are merged and a new client
version is released, `builld.sbt` can be updated to the new version.

* https://github.com/guardian/anghammarad/pull/33
* https://github.com/guardian/anghammarad/pull/34