Create DNS validated certificate within the cloudformation template.
With https://github.com/guardian/dns-validation-lambda it is now possible to create DNS validated certificates that require no human interaction to be verified.

DNS validate certificates are preferred over email validated ones as they auto-renew; email validated certificates require a manual step of clicking a link in an email.

This changes the template to accept the domain name as a parameter, which is used to create the certificate.