Move cloudwatch code to its own MetricService
The idea is for cloudwatch to pick up the latest data in cache once
every 6 hours.

Given that we're working with multiple streams of time series data
that we want to add together, the code has a guard clause to skip
posting to cloudwatch if the data set is incomplete.

This is a guarantee that every data stream has the same number of
datapoints in the same time range, which facilitates plotting graphs
like "total number of vulnerabilities for account frontend over
time".