Log GCP vulnerabilities to cloudwatch metrics
This only includes HIGHs and CRITICALs.

The fact that GCP vulnerabilities have a project, as opposed to AWS
vulnerabilities which have an account, means that their respective
cloudwatch metrics will have different dimensions, which means they
will have different schemas.

Accounting for this requires a refactor of putMetric into something
parameterised that can be used for both cases.