add dynamoDB and put notification data
This PR adds dynamoDB so that Security HQ's IAM notification
job can record which notifications have been successfully
sent. The aim is that this data will be used for two reasons.
First, so that a new feature, the Credentials Reaper, can be
built which automatically disables permanent credentials which
are either deemed too old or are missing multi factor
authentication. This new feature should only disable a credential
if 2 notifications have been sent, hence the need to store state
in a database.
The second reason is that we hope this data will provide a
helpful audit trail for Infosec.