Merge two vulnerable user types into one
I was struggling to write the credentials reaper feature managing two
separate user case classes, one for keys which are old and the other for
keys that are missing mfa.

This change makes it easier to complete the credentials reaper feature,
however the email alert to users will not include why the user has
been flagged. This shouldn't be a major problem as these details
are including on the security HQ dashboard.