add endpoint to get historical iam alerts
This PR adds a new route, which returns all of the data in the
IAM dynamoDB table. This is useful because a) it's helped us
test that PR#256 works! and b) it means that any user can
see which alerts have been sent and what the deadlines are
on any flagged credentials.