Adds IAM UpdateAccessKey permission
Currently Security HQ sends AWS accounts an email alert when they
have a user with a permanent credential which needs rotating or lacks
multi-factor authentication.

The next step for this feature is to create what we call the
Credentials Reaper, which disables access keys if they have not been
rectified after being notified.

In order to do that, SHQ needs IAM permissions to make an
access key inactive, which is what this PR does.