Turn off the IAM job
This change stops the IAM job from running, which means that no
notifications about permanent credentials will be sent to colleagues.
However, the test endpoint /iam/test-notifications, still works so
we will be able to continue testing this feature.
When the credentials reaper feature is ready to be released to PROD
the line in app components can be uncommented.