Refactor IAM job logic to make it easier to read
A lot of the IAM notifications job logic was in one file, which was
really difficult to read and make changes to. I got tired of scrolling
down a long file to find the logic I was searching for. So, I've
separated the code out into separate files to make it easier to read
and work with. No logic has actually been changed. Just a little
spring clean 🧚🏾‍♂️