Fixes missing deadline field in IAM email alerts
The problem was that the IAM notification emails sent by security HQ
(to alert users that there is a vulnerable permanent credential in their
account), were missing the deadline field.

This is an important problem, because it means that readers of the email
would not be sure when security hq was going to a automatically delete
the problematic permanent credential.

This PR fixes this by ensuring that the logic which creates the email
messages uses an existing function that creates a deadline if it does
not already exist.

This may be the case if the credential is being alerted for the first
time and is therefore not present in dynamoDB.