Improve formatting of IAM notification emails
Security HQ sends email notifications when IAM permanent credentials
are vulnerable. Currently the email formatting is a little bit
confusing, because there's a lot of text. As a quick fix, this PR
adds a line of dashes, which gmail formats quite nicely as a line
that runs across the email dividing the part of the email that
details the flagged users and the part of the email that
describes where to go for more information.

Hopefully this quick fix helps the readability of the email!