Remove access key and password for vulnerable users
IAM users with permanent credentials present a security vulnerability
to the Guardian, however sometimes they are required. In this case,
the credentials should be rotated regularly to minimise the security
risk of them being compromised.

This PR removes the access key and password for users with permanent
credentials who have been notified twice that these credentials
require rotiation, but have taken no action.

Once this has been done, a final email notification will be sent to
inform the AWS account owners that this has been done.