Find out when user's key was last rotated
Security HQ aims to automatically delete permanent credentials, which
are due to be rotated but have not been. In order to disable a user's
access key, AWS requires an access key id. Each user can have two
access keys, so in order to determine which access key should be
disabled, this change adds a field to each vulnerable user which
knows when the access key was last rotated.

I'm hoping that by having access to the last rotation date of the
access key, SHQ will be able to determine which of the access keys
for a given user should be deleted.