Refactor code which determines vulnerable users
Prior to this change, one function was doing the work of both finding
vulnerable users and deciding which to alert. Not all vulnerable users
require alerting at a particular time. This depends on whether they
have already received a notification or not. So, this change splits out
that logic. This is beneficial, because it means that the code is
more reusable and we can use the same function for the automatic
disabling of credentials.