Small fixes to the reaper after testing
There were a couple of problems with the reaper, which testing has
surfaced. This PR fixes the following:

1) the AWS API update-access-key requires a username when the
credentials used to call the API are not the same as the key that
is being updated.

This has been fixed by adding a username to the API call.

2) the usersToDisable function in IamJob creates two lists of users to
disable.

The first list are users who have old access keys and the second
is a list of users who require multi-factor authentication. In our
test case, the vulnerable user had both old access keys and required
mfa.

This meant that the user was present in both lists, resulting in
two API calls to remove the password and delete the access key of the
same user. In addition, this meant that the email to AWS accounts
included this user twice, which is confusing.

This has been fixed by ensuring that the list of vulnerable users
is of uniques only.