Turn on the Credentials Reaper for beta testers
This PR turns on the Credentials Reaper job for beta tester AWS
accounts; frontend, mobile and deploy tools. This will be the first
time that the reaper is in production with its ability to disable
active access keys and remove passwords, so we're just testing it
with a few accounts so as not to wreak havoc across the department!