Gets Janus Data file in S3 & refactors job
Part 1 - getting Janus Data file from S3
Unrecognised staff are verified by comparing Janus users
to the tags of those with IAM permanent credentials.

If any of the tags (which will have matching usernames to the ones
in Janus) are present amongst the IAM users, which are NOT in the
list of Janus users, that IAM cred will be disabled.

In order for the above to work, Security HQ needs to read a file
in S3 which contains a list of Janus users.

Part 2 - refactoring the unrecognised job
The IAM unrecognised job has been refactored to use a for
comprehension at the edge of the programme that contains
all of the steps that the unrecognised job does.

This is a pattern that Security HQ uses elsewhere and it
makes the logic easier to follow.

The IAM vulnerable user job does not follow this pattern, so
it would be good to refactor this also in the future.