Create separate dynamoDB table for DEV mode
Currently, the PROD dynamo table is used for testing purposes, but
this is not ideal. This PR creates a DEV table to be used for this
instead.

This means that developers working on SHQ will need to change the
value of IAM_DYNAMO_TABLE_NAME in their local conf file.

Also this PR deletes an unused bash script related to an old
local dynamoDB implementation.