Dashboard shows which human users have tag
The unrecognised user feature will automatically remove permanent
IAM users who are not in janus.

For this to work, each human credential must be tagged. This PR adds
a column to the IAM table showing if the human user has the correct
tag.

Also, the status of the user is now Amber for credentials which
comply with all other rules, but do not have the required tag.