CDK full (#329)
Adds a fuller CDK stack. A follow on PR will be raised to remove
the old stack and migrate the shared resources.

Note, for the London CIDR, ideally we would read from
https://github.com/guardian/private-infrastructure-config here,
but we don't yet have a good solution for the access token to allow
GHA to read from this (private) repo. So will update once we do.

Co-authored-by: Jacob Winch <jacob.winch@guardian.co.uk>