Upgrade SBT to 1.5.7 to protect from log4j exploit
Due to the log4j exploit, this PR upgrades sbt to 1.5.7 where log4j
has been updated to a safe version. This means that it should not be
possible for SHQ to be compromised using log4j in sbt. This possibility
is very much an edge case, but this change seems prudent nevertheless.

The sbt upgrade also required a play upgrade, which triggered binary
incompatibility warnings with scala-java8-compat. This has been fixed
using a `libraryDependencySchemes` setting as described here:
https://github.com/scanamo/scanamo/pull/1285.