Avoid unnecessary (mutually exclusive) setter usage
Log shipping via kinesis was updated to use AWS SDK v2 (via the kinesis logback appender v2), but that included a change in the SDK that validated the existence of an STS client when `.setRoleToAssumeArn(role)` is called. This setter is unnecessary, as `setCredentialsProvider` is preferred for our usage, which appears to now be (strictly) mutually exclusive.