Make Snyk reports individually addressable.
This change adds a snykOrg view and controller function that allows requests to `/snyk/:snykOrk` to be routed to individual organisation pages.

The motivation for this change is to allow teams to be given links to their own page on Security HQ to provide a more focussed view of their issues.