refactor: Remove root `package.json` file
The `package.json` file at the root enables prettier and eslint on the JavaScript in the repository.
It also does some Snyk checks, which have been superseded by GitHub Actions.

Whilst this is good, we're not checking these as part of the build.
Additionally, we don't have a standard client-side setup in this repository, with `jquery` and `materialize` committed to VCS.

That is, in it's current state, the root `package.json` isn't providing much utility in this repository, so remove it.