Improve error handling / logging
Log messages on Failures now include a summary of the underlying
cause, if present. This should improve error visibility across the
board in Security HQ.

We also add additional error handling to the cache service. On every
cache update we include a summary of which accounts have failures in
their cache data.